+++ src/main.ts
@@ -1,6 +1,7 @@
 import { join } from "https://deno.land/std@0.217.0/path/join.ts";
 
 import plist from "npm:plist";
+import { flavorEntries, type FlavorName } from "npm:@catppuccin/palette@1.1.0";
 import Latte from "npm:@catppuccin/vscode/themes/latte.json" with { type: "json" };
 import Frappe from "npm:@catppuccin/vscode/themes/frappe.json" with { type: "json" };
 import Macchiato from "npm:@catppuccin/vscode/themes/macchiato.json" with { type: "json" };
@@ -10,33 +11,38 @@ import { convert } from "./convert.ts";
 
 export type VSCTheme = typeof Latte;
 
-const themes = {
-  Latte: {
+const themes: Record<FlavorName, { uuid: string; vscode: VSCTheme }> = {
+  latte: {
     uuid: "96a262cd-4b2f-49f5-9125-8dd0077cbfe1",
     vscode: Latte satisfies VSCTheme,
   },
-  Frappe: {
+  frappe: {
     uuid: "e0ada983-8938-490c-86f0-97a1a0ec58e4",
     vscode: Frappe satisfies VSCTheme,
   },
-  Macchiato: {
+  macchiato: {
     uuid: "02b2bdf3-9eb7-4396-bf04-f17f1468f99f",
     vscode: Macchiato satisfies VSCTheme,
   },
-  Mocha: {
+  mocha: {
     uuid: "627ce890-fabb-4d39-9819-7be71f4bdca7",
     vscode: Mocha satisfies VSCTheme,
   },
 };
 
 const outDir = new URL("../themes", import.meta.url).pathname;
-
 Deno.mkdirSync(outDir, { recursive: true });
 
-Object.entries(themes).forEach(([name, { uuid, vscode }]) => {
-  const plistContent = plist.build(convert(name.toLocaleLowerCase() as FlavorName, vscode, uuid));
-  const fileName = `Catppuccin ${name}.tmTheme`;
+const capitalize = (s: string) => s.charAt(0).toUpperCase() + s.slice(1);
 
-  Deno.writeTextFile(join(outDir, fileName), plistContent);
-  console.log(`Wrote ${fileName}`);
-});
+if (import.meta.main) {
+  flavorEntries.map(([flavorName]) => {
+    const { uuid, vscode } = themes[flavorName];
+
+    const plistContent = plist.build(convert(flavorName, vscode, uuid));
+    const fileName = `Catppuccin ${capitalize(flavorName)}.tmTheme`;
+
+    Deno.writeTextFile(join(outDir, fileName), plistContent);
+    console.log(`Wrote ${fileName}`);
+  });
+}
